backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
# openai==1.3.7  # Replaced with Gemini API
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
